correction des migration pour lassignement rapide

+++ database/migrations/2026_07_14_000002_create_order_broadcasts_table.php
@@ -10,13 +10,13 @@ return new class extends Migration
     {
         Schema::create('order_broadcasts', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('order_id');
+            $table->unsignedInteger('order_id');
             $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
             $table->decimal('radius_km', 5, 1)->default(5.0);
             $table->timestamp('started_at')->useCurrent();
             $table->timestamp('expires_at')->nullable();
             $table->enum('status', ['pending', 'assigned', 'failed', 'expired'])->default('pending');
-            $table->unsignedBigInteger('assigned_to')->nullable();
+            $table->unsignedInteger('assigned_to')->nullable();
             $table->foreign('assigned_to')->references('id')->on('users')->onDelete('set null');
             $table->timestamp('assigned_at')->nullable();
             $table->timestamps();
